backend/requirements.txt
# DeepSeek ReAct 智能体 - Python 依赖包

# HTTP 请求库 - 用于调用 DeepSeek API
requests>=2.31.0,<3.0.0

# 环境变量管理 - 用于加载 .env 文件中的配置
python-dotenv>=1.0.0,<2.0.0

# 彩色终端输出（可选但推荐）- 用于美化 CLI 界面
colorama>=0.4.6,<1.0.0

# 系统进程和端口管理
psutil>=5.9.0,<6.0.0

# FastAPI - Web 框架
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
websockets>=12.0,<13.0

# Pydantic - 数据验证
pydantic>=2.0.0,<3.0.0

# Pillow - 图片处理（用于截图和 GIF 生成）
pillow>=10.0.0,<11.0.0

# 文档解析库
PyPDF2>=3.0.0,<4.0.0
python-docx>=1.0.0,<2.0.0

